requirements.txt
ipykernel
langchain
python-dotenv
langchain-google-genai
langchain_google_genai
beautifulsoup4
langchain_community
faiss-cpu
fastapi
streamlit
uvicorn
transformers
bs4
google-generativeai
langchain-google-genai
langchain-groq
langchain-chroma
arxiv
wikipedia
sentence-transformers
mysql-connector-python
SQLAlchemy
pypdf
validators
youtube_transcript_api
unstructured
pytube
numexpr
langchain_huggingface
huggingface_hub
gradio
langchain_nvidia_ai_endpoints
faiss-cpu